add select alpha version
