+++ src/Select/Select.tsx
@@ -1,127 +1,102 @@
-import React, { useEffect, useRef, useCallback } from 'react';
-import { Input } from '../Input';
+import React, { useReducer, useRef, createContext, useContext, useMemo } from 'react';
+import { useTheme } from '../Themes';
 import { ArrowDownIcon } from './icons/ArrowDownIcon';
-import { Dropdown, useDropdown } from '../Dropdown';
-import { OptionsList } from './OptionsList';
-import { useSelected } from './hooks';
-import { ISelectProps } from './types';
+import { useSelectEvents } from './use-select';
+import { ISelectProps, ISelectState, SelectChangeTypes, ISelectContext } from './types';
+import { SelectWrapper } from './styles';
+
+const initState = {
+  type: SelectChangeTypes.idle,
+  isOpen: false,
+};
+
+const selectReducer = (state: ISelectState, changes: ISelectState) => {
+  switch (changes.type) {
+    case SelectChangeTypes.idle:
+    default:
+      return {
+        ...changes,
+      };
+    case SelectChangeTypes.selectClick:
+      return {
+        ...state,
+        isOpen: !state.isOpen,
+      };
+    case SelectChangeTypes.clickOutside:
+    case SelectChangeTypes.scroll:
+      return {
+        ...state,
+        isOpen: false,
+      };
+    case SelectChangeTypes.changeDisplayValue:
+      return {
+        ...state,
+        ...changes,
+        isOpen: false,
+      };
+  }
+};
+
+const SelectContext = createContext<ISelectContext | undefined>(undefined);
 
 export function Select(props: ISelectProps) {
-    const {
-        name,
-        options,
-        onChangeHandler,
-        disabled,
-        value,
-        hasError,
-        onBlurHandler,
-        onFocusHandler,
-        dropdownStyles,
-        styles,
-        children,
-        ...restProps
-    } = props;
-    const selectRef = useRef<HTMLInputElement>(null);
-    const dropdownRef = useRef<HTMLDivElement>(null);
+  const {
+    stateReducer = selectReducer,
+    value,
+    caption,
+    prefix,
+    prefixCss,
+    isEdit,
+    inputCss,
+    displayValueCss,
+    children,
+    ...restProps
+  } = props;
 
-    const { selectedValue, valueSetter, resetValue, caption } = useSelected(options);
-    const { isOpen, parentRect, setIsOpen } = useDropdown(selectRef, dropdownRef);
+  const theme = useTheme();
 
-    const handleKeypress = useCallback(
-        (event: KeyboardEvent) => {
-            if (event.keyCode === 32) {
-                setIsOpen(!isOpen);
-            }
-        },
-        [isOpen, setIsOpen]
-    );
+  const dropdownRef = useRef<HTMLDivElement>(null);
+  const selectRef = useRef<HTMLDivElement>(null);
+  const selectRect = selectRef?.current?.getBoundingClientRect();
 
-    useEffect(() => {
-        if (value && options && Object.keys(options).length > 0) {
-            valueSetter(value);
-        } else if (!value) {
-            resetValue();
-        }
-        if (isOpen) {
-            selectRef?.current?.focus();
-        }
-    }, [value, valueSetter, options, resetValue, isOpen]);
+  const [{ isOpen }, dispatch] = useReducer(stateReducer, initState);
+  useSelectEvents(selectRef, dropdownRef, dispatch, { isOpen });
 
-    const handleClick: React.MouseEventHandler<HTMLDivElement> = (event) => {
-        event.preventDefault();
-        setIsOpen(!isOpen);
-    };
+  const handleClick = () => {
+    dispatch({ type: SelectChangeTypes.selectClick });
+  };
 
-    const handleItemClick = useCallback(
-        (optionValue: string) => {
-            valueSetter(optionValue);
-            selectRef.current.focus();
-            onChangeHandler(optionValue);
-            setIsOpen(false);
-        },
-        [onChangeHandler, setIsOpen, valueSetter]
-    );
+  const ctxValue = useMemo<ISelectContext>(() => ({ dropdownRef: dropdownRef, isOpen, selectRect }), [
+    isOpen,
+    selectRect,
+  ]);
 
-    const handleFocus = () => {
-        const select = selectRef.current;
-        select.addEventListener('keypress', handleKeypress);
-        onFocusHandler && onFocusHandler(selectedValue);
-    };
-
-    const handleBlur = () => {
-        const select = selectRef.current;
-        select.removeEventListener('keypress', handleKeypress);
-        onBlurHandler && onBlurHandler(selectedValue);
-    };
-
-    const renderOptions = useCallback(() => {
-        if (children) {
-            return children({ selectedValue, changeHandler: onChangeHandler });
-        } else {
-            return (
-                <OptionsList
-                    options={options}
-                    selectedValue={selectedValue}
-                    onChangeHandler={handleItemClick}
-                />
-            );
-        }
-    }, [children, handleItemClick, onChangeHandler, options, selectedValue]);
-
-    return (
-        <div
-            css={{
-                position: 'relative',
-                maxWidth: 'auto'
-            }}>
-            <Input
-                name={name}
-                onChangeHandler={handleItemClick}
-                ref={selectRef}
-                title={caption}
-                {...restProps}
-                onClick={disabled ? undefined : handleClick}
-                onFocusHandler={handleFocus}
-                onBlurHandler={handleBlur}
-                value={caption}
-                css={{ borderRadius: '4px', padding: '0.5rem', width: '100%', height: 'auto', ...styles }}
-                // style={{ padding: 0, ...styles }} // for input wrapper
-                readOnly
-                hasError={hasError}
-            />
-            <ArrowDownIcon
-                css={{
-                    position: 'absolute',
-                    top: '30%',
-                    right: '10px',
-                    transform: `rotate(${isOpen ? '180deg' : '0'})`,
-                    transition: 'all 0.2s ease-in-out',
-                }}
-            />
-
-            <Dropdown isOpen={isOpen} parentBound={parentRect} ref={dropdownRef} styles={dropdownStyles}>
-                {renderOptions()}
-            </Dropdown>
-        </div>
-    );
+  return (
+    <SelectWrapper
+      onClick={handleClick}
+      ref={selectRef}
+      css={{
+        borderColor: `${isOpen ? theme.colors.mainColors.primary : theme.colors.pageElementsColors.border}`,
+      }}
+      {...restProps}
+    >
+      <SelectContext.Provider value={ctxValue}>{children}</SelectContext.Provider>
+      <ArrowDownIcon
+        fill={theme.colors.textColors.text}
+        css={{
+          marginLeft: 'auto',
+          transform: `rotate(${isOpen ? '180deg' : '0'})`,
+          transition: 'all 0.3s ease-in-out',
+        }}
+      />
+    </SelectWrapper>
+  );
 }
+
+export const useSelect = () => {
+  const context = useContext(SelectContext);
+  if (!context) {
+    throw new Error('Component must be inside Select ');
+  }
+  return context;
+};
