--- src/Select2/index.ts
@@ -1,3 +1,0 @@
-export * from './Select'
-export * from './SelectOptions';
-export * from './types';
