+++ src/index.ts
@@ -7,6 +7,6 @@ export * from './Checkbox';
 export * from './InputNumber';
 export * from './Portal';
 export * from './Dropdown';
-export * from './Select2';
+export * from './Select';
 export * from './Modal';
 // export * from './DayPicker';
